refactored detail view
